+++ Lesson8/src/main/java/ru/utils/ProductFilter.java
@@ -4,14 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter @Setter
 @AllArgsConstructor
 public class ProductFilter {
 
-    private BigDecimal minPrice;
-    private BigDecimal maxPrice;
+    private Float minPrice;
+    private Float maxPrice;
     private String partName;
 
 }
